requirements.txt
discord.py
requests
bs4
aiohttp
websockets
chardet
colorama
random2
pyautogui
datetime
gTTS
